+++ code/app/src/main/java/com/example/bread/fragment/HomeFragment.java
@@ -1,66 +1,37 @@
 package com.example.bread.fragment;
 
-import android.app.AlertDialog;
+import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
+import android.widget.ListView;
 import android.widget.Toast;
 
 import androidx.fragment.app.Fragment;
 
 import com.example.bread.R;
+import com.example.bread.controller.HistoryMoodEventArrayAdapter;
 import com.example.bread.model.MoodEvent;
-import com.example.bread.model.MoodEvent.SocialSituation;
+import com.example.bread.repository.MoodEventRepository;
+import com.example.bread.view.LoginPage;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 
-import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link HomeFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
 public class HomeFragment extends Fragment {
 
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
+    private static final String TAG = "HomeFragment";
+    private ArrayList<MoodEvent> moodEventArrayList;
+    private HistoryMoodEventArrayAdapter moodEventArrayAdapter;
+    private MoodEventRepository moodEventRepository;
+    private FirebaseAuth mAuth;
 
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
 
     public HomeFragment() {
         // Required empty public constructor
-    }
-
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment HomeFragment.
-     */
-    // TODO: Rename and change types and number of parameters
-    public static HomeFragment newInstance(String param1, String param2) {
-        HomeFragment fragment = new HomeFragment();
-        Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
     }
 
     @Override
@@ -68,59 +39,40 @@ public class HomeFragment extends Fragment {
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
         View view = inflater.inflate(R.layout.fragment_home, container, false);
+        ListView moodEventListView = view.findViewById(R.id.homeListView);
+        moodEventArrayList = new ArrayList<>();
+        moodEventArrayAdapter = new HistoryMoodEventArrayAdapter(getContext(), moodEventArrayList);
+        moodEventListView.setAdapter(moodEventArrayAdapter);
+
+        mAuth = FirebaseAuth.getInstance();
+        moodEventRepository = new MoodEventRepository();
 
         // Note: To use clicking functionality, when you implement the ListView and adapter later,
         // you'll need to add this line:
         // moodArrayAdapter.setOnMoodEventClickListener(this::showMoodDetailsDialog);
-
+        fetchMoodEvents();
         return view;
     }
 
-    /**
-     * Shows a dialog with the details of the selected mood event.
-     *
-     * @param moodEvent The mood event to show details for
-     */
-    public void showMoodDetailsDialog(MoodEvent moodEvent) {
-        if (getContext() == null) return;
-
-        AlertDialog.Builder builder = new AlertDialog.Builder(getContext());
-        builder.setTitle("View Mood");
-
-        // Inflate a custom layout for the dialog
-        View dialogView = getLayoutInflater().inflate(R.layout.dialog_mood_details, null);
-
-        // Set up the views
-        TextView emotionTextView = dialogView.findViewById(R.id.detail_emotion);
-        TextView dateTextView = dialogView.findViewById(R.id.detail_date);
-        TextView reasonTextView = dialogView.findViewById(R.id.detail_reason);
-        TextView socialSituationTextView = dialogView.findViewById(R.id.detail_social_situation);
-
-        // Set the data
-        emotionTextView.setText(moodEvent.getEmotionalState().toString());
-
-        // Format date
-        SimpleDateFormat formatter = new SimpleDateFormat("dd MMM yyyy hh:mm a");
-        String dateString = formatter.format(moodEvent.getTimestamp());
-        dateTextView.setText(dateString);
-
-        // Set reason
-        reasonTextView.setText(moodEvent.getReason() != null ? moodEvent.getReason() : "No reason provided");
-
-        // Set social situation
-        SocialSituation situation = moodEvent.getSocialSituation();
-        socialSituationTextView.setText(situation != null ? situation.toString() : "Not specified");
-
-        builder.setView(dialogView);
-        builder.setPositiveButton("Close", (dialog, which) -> dialog.dismiss());
-
-        // Add an Edit button
-        builder.setNeutralButton("Edit", (dialog, which) -> {
-            // TODO: Implement edit functionality in future updates
-            Toast.makeText(getContext(), "Edit functionality to be implemented", Toast.LENGTH_SHORT).show();
-        });
-
-        AlertDialog dialog = builder.create();
-        dialog.show();
+    private void fetchMoodEvents() {
+        FirebaseUser user = mAuth.getCurrentUser();
+        if (user != null) {
+            String username = user.getDisplayName();
+            if (username != null) {
+                moodEventRepository.listenForEventsFromFollowing(username, moodEvents -> {
+                    moodEventArrayList.clear();
+                    moodEventArrayList.addAll(moodEvents);
+                    moodEventArrayAdapter.notifyDataSetChanged();
+                }, e -> {
+                    Log.e(TAG, "Failed to fetch mood events for user: " + username, e);
+                    Toast.makeText(getContext(), "Failed to fetch mood events", Toast.LENGTH_SHORT).show();
+                });
+            }
+        } else {
+            Log.e(TAG, "User is not logged in");
+            Intent intent = new Intent(getContext(), LoginPage.class);
+            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+            startActivity(intent);
+        }
     }
 }
